Refactor ops and props
