+++ scripts/addons/cam/bas_relief.py
@@ -1,4 +1,4 @@
-"""Fabex 'basrelief.py'
+"""Fabex 'bas_relief.py'
 
 Module to allow the creation of reliefs from Images or View Layers.
 (https://en.wikipedia.org/wiki/Relief#Bas-relief_or_low_relief)
@@ -12,18 +12,11 @@ import numpy
 
 import bpy
 
-from bpy.types import PropertyGroup
-from bpy.props import (
-    BoolProperty,
-    EnumProperty,
-    FloatProperty,
-    IntProperty,
-    PointerProperty,
-    StringProperty,
+from .constants import EPS, NUMPYALG
+from .utilities.image_utils import (
+    image_to_numpy,
+    numpy_to_image,
 )
-
-from .constants import PRECISION, EPS, NUMPYALG
-from .utilities.image_utils import image_to_numpy, numpy_save, numpy_to_image
 
 
 class ReliefError(Exception):
@@ -653,140 +646,6 @@ def linear_bcg(n, b, x, itol, tol, itmax, iter, err, rows, cols, planar):
             break
 
 
-# def numpy_save(a, iname):
-#     """Save a NumPy array as an image file in OpenEXR format.
-
-#     This function takes a NumPy array and saves it as an image file using
-#     Blender's rendering capabilities. It configures the image settings to
-#     use the OpenEXR format with black and white color mode and a color depth
-#     of 32 bits. The rendered image is saved to the specified filename.
-
-#     Args:
-#         a (numpy.ndarray): The NumPy array to be saved as an image.
-#         iname (str): The filename (including path) where the image will be saved.
-#     """
-
-#     inamebase = bpy.path.basename(iname)
-
-#     i = numpy_to_image(a, inamebase)
-
-#     r = bpy.context.scene.render
-
-#     r.image_settings.file_format = "OPEN_EXR"
-#     r.image_settings.color_mode = "BW"
-#     r.image_settings.color_depth = "32"
-
-#     i.save_render(iname)
-
-
-# def numpy_to_image(a, iname):
-#     """Convert a NumPy array to a Blender image.
-
-#     This function takes a NumPy array and converts it into a Blender image.
-#     It first checks if an image with the specified name and dimensions
-#     already exists in Blender. If it does, that image is used; otherwise, a
-#     new image is created with the specified name and dimensions. The
-#     function then reshapes the NumPy array to match the image format and
-#     assigns the pixel data to the image.
-
-#     Args:
-#         a (numpy.ndarray): A 2D NumPy array representing the pixel data of the image.
-#         iname (str): The name to assign to the Blender image.
-
-#     Returns:
-#         bpy.types.Image: The Blender image created or modified with the pixel data from the NumPy
-#             array.
-#     """
-
-#     t = time.time()
-#     print("Numpy to Image - Here")
-#     t = time.time()
-#     print(a.shape[0], a.shape[1])
-#     foundimage = False
-#     for image in bpy.data.images:
-
-#         if (
-#             image.name[: len(iname)] == iname
-#             and image.size[0] == a.shape[0]
-#             and image.size[1] == a.shape[1]
-#         ):
-#             i = image
-#             foundimage = True
-#     if not foundimage:
-#         bpy.ops.image.new(
-#             name=iname,
-#             width=a.shape[0],
-#             height=a.shape[1],
-#             color=(0, 0, 0, 1),
-#             alpha=True,
-#             generated_type="BLANK",
-#             float=True,
-#         )
-#         for image in bpy.data.images:
-
-#             if (
-#                 image.name[: len(iname)] == iname
-#                 and image.size[0] == a.shape[0]
-#                 and image.size[1] == a.shape[1]
-#             ):
-#                 i = image
-
-#     d = a.shape[0] * a.shape[1]
-#     a = a.swapaxes(0, 1)
-#     a = a.reshape(d)
-#     a = a.repeat(4)
-#     a[3::4] = 1
-#     # i.pixels=a
-#     i.pixels[:] = a[:]  # this gives big speedup!
-#     print("\ntime " + str(time.time() - t))
-#     return i
-
-
-# def image_to_numpy(i):
-#     """Convert an image to a NumPy array.
-
-#     This function takes an image object and converts its pixel data into a
-#     NumPy array. It first retrieves the pixel data from the image, then
-#     reshapes and rearranges it to match the image's dimensions. The
-#     resulting array is structured such that the height and width of the
-#     image are preserved, and the color channels are appropriately ordered.
-
-#     Args:
-#         i (Image): An image object that contains pixel data.
-
-#     Returns:
-#         numpy.ndarray: A 2D NumPy array representing the pixel data of the image.
-
-#     Note:
-#         The function optimizes performance by directly accessing pixel data
-#         instead of using slower methods.
-#     """
-
-#     t = time.time()
-#     inc = 0
-
-#     width = i.size[0]
-#     height = i.size[1]
-#     x = 0
-#     y = 0
-#     count = 0
-#     na = numpy.array((0.1), dtype=numpy.float64)
-
-#     size = width * height
-#     na.resize(size * 4)
-
-#     # these 2 lines are about 15% faster than na=i.pixels[:].... whyyyyyyyy!!?!?!?!?! Blender image data access is evil.
-#     p = i.pixels[:]
-#     na[:] = p
-#     # na=numpy.array(i.pixels[:])#this was terribly slow... at least I know why now, it probably
-#     na = na[::4]
-#     na = na.reshape(height, width)
-#     na = na.swapaxes(0, 1)
-
-#     print("\ntime of image to numpy " + str(time.time() - t))
-#     return na
-
-
 def tonemap(i, exponent):
     """Apply tone mapping to an image array.
 
@@ -1291,382 +1150,3 @@ def relief(br):
 
     t = time.time() - t
     print("total time:" + str(t) + "\n")
-
-
-class DoBasRelief(bpy.types.Operator):
-    """Calculate Bas Relief"""
-
-    bl_idname = "scene.calculate_bas_relief"
-    bl_label = "Calculate Bas Relief"
-    bl_options = {"REGISTER", "UNDO"}
-
-    processes = []
-
-    # Bas Relief Properties
-    use_image_source: BoolProperty(
-        name="Use Image Source",
-        description="",
-        default=False,
-    )
-    source_image_name: StringProperty(
-        name="Image Source",
-        description="image source",
-    )
-    view_layer_name: StringProperty(
-        name="View Layer Source",
-        description="Make a bas-relief from whatever is on this view layer",
-    )
-    bit_diameter: FloatProperty(
-        name="Diameter of Ball End in mm",
-        description="Diameter of bit which will be used for carving",
-        min=0.01,
-        max=50.0,
-        default=3.175,
-        precision=PRECISION,
-    )
-    pass_per_radius: IntProperty(
-        name="Passes per Radius",
-        description="Amount of passes per radius\n(more passes, " "more mesh precision)",
-        default=2,
-        min=1,
-        max=10,
-    )
-    width_mm: IntProperty(
-        name="Desired Width in mm",
-        default=200,
-        min=5,
-        max=4000,
-    )
-    height_mm: IntProperty(
-        name="Desired Height in mm",
-        default=150,
-        min=5,
-        max=4000,
-    )
-    thickness_mm: IntProperty(
-        name="Thickness in mm",
-        default=15,
-        min=5,
-        max=100,
-    )
-
-    justify_x: EnumProperty(
-        name="X",
-        items=[("1", "Left", "", 0), ("-0.5", "Centered", "", 1), ("-1", "Right", "", 2)],
-        default="-1",
-    )
-    justify_y: EnumProperty(
-        name="Y",
-        items=[
-            ("1", "Bottom", "", 0),
-            ("-0.5", "Centered", "", 2),
-            ("-1", "Top", "", 1),
-        ],
-        default="-1",
-    )
-    justify_z: EnumProperty(
-        name="Z",
-        items=[
-            ("-1", "Below 0", "", 0),
-            ("-0.5", "Centered", "", 2),
-            ("1", "Above 0", "", 1),
-        ],
-        default="-1",
-    )
-
-    depth_exponent: FloatProperty(
-        name="Depth Exponent",
-        description="Initial depth map is taken to this power. Higher = " "sharper relief",
-        min=0.5,
-        max=10.0,
-        default=1.0,
-        precision=PRECISION,
-    )
-
-    silhouette_threshold: FloatProperty(
-        name="Silhouette Threshold",
-        description="Silhouette threshold",
-        min=0.000001,
-        max=1.0,
-        default=0.003,
-        precision=PRECISION,
-    )
-    recover_silhouettes: BoolProperty(
-        name="Recover Silhouettes",
-        description="",
-        default=True,
-    )
-    silhouette_scale: FloatProperty(
-        name="Silhouette Scale",
-        description="Silhouette scale",
-        min=0.000001,
-        max=5.0,
-        default=0.3,
-        precision=PRECISION,
-    )
-    silhouette_exponent: IntProperty(
-        name="Silhouette Square Exponent",
-        description="If lower, true depth distances between objects will be "
-        "more visibe in the relief",
-        default=3,
-        min=0,
-        max=5,
-    )
-    attenuation: FloatProperty(
-        name="Gradient Attenuation",
-        description="Gradient attenuation",
-        min=0.000001,
-        max=100.0,
-        default=1.0,
-        precision=PRECISION,
-    )
-    min_gridsize: IntProperty(
-        name="Minimum Grid Size",
-        default=16,
-        min=2,
-        max=512,
-    )
-    smooth_iterations: IntProperty(
-        name="Smooth Iterations",
-        default=1,
-        min=1,
-        max=64,
-    )
-    vcycle_iterations: IntProperty(
-        name="V-Cycle Iterations",
-        description="Set higher for planar constraint",
-        default=2,
-        min=1,
-        max=128,
-    )
-    linbcg_iterations: IntProperty(
-        name="LINBCG Iterations",
-        description="Set lower for flatter relief, and when using " "planar constraint",
-        default=5,
-        min=1,
-        max=64,
-    )
-    use_planar: BoolProperty(
-        name="Use Planar Constraint",
-        description="",
-        default=False,
-    )
-    gradient_scaling_mask_use: BoolProperty(
-        name="Scale Gradients with Mask",
-        description="",
-        default=False,
-    )
-    decimate_ratio: FloatProperty(
-        name="Decimate Ratio",
-        description="Simplify the mesh using the Decimate modifier. "
-        "The lower the value the more simplyfied",
-        min=0.01,
-        max=1.0,
-        default=0.1,
-        precision=PRECISION,
-    )
-
-    gradient_scaling_mask_name: StringProperty(
-        name="Scaling Mask Name",
-        description="Mask name",
-    )
-    scale_down_before_use: BoolProperty(
-        name="Scale Down Image Before Processing",
-        description="",
-        default=False,
-    )
-    scale_down_before: FloatProperty(
-        name="Image Scale",
-        description="Image scale",
-        min=0.025,
-        max=1.0,
-        default=0.5,
-        precision=PRECISION,
-    )
-    detail_enhancement_use: BoolProperty(
-        name="Enhance Details",
-        description="Enhance details by frequency analysis",
-        default=False,
-    )
-    # detail_enhancement_freq=FloatProperty(name="frequency limit", description="Image scale", min=0.025, max=1.0, default=.5, precision=PRECISION)
-    detail_enhancement_amount: FloatProperty(
-        name="Amount",
-        description="Image scale",
-        min=0.025,
-        max=1.0,
-        default=0.5,
-        precision=PRECISION,
-    )
-
-    advanced: BoolProperty(
-        name="Advanced Options",
-        description="Show advanced options",
-        default=True,
-    )
-
-    def invoke(self, context, event):
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self)
-
-    def execute(self, context):
-        """Execute the relief rendering process based on the provided context.
-
-        This function retrieves the scene and its associated bas relief
-        settings. It checks if an image source is being used and sets the view
-        layer name accordingly. The function then attempts to render the scene
-        and generate the relief. If any errors occur during these processes,
-        they are reported, and the operation is canceled.
-
-        Args:
-            context: The context in which the function is executed.
-
-        Returns:
-            dict: A dictionary indicating the result of the operation, either
-        """
-
-        if not self.use_image_source and self.view_layer_name == "":
-            self.view_layer_name = bpy.context.view_layer.name
-
-        try:
-            render_scene(
-                self.width_mm,
-                self.height_mm,
-                self.bit_diameter,
-                self.pass_per_radius,
-                not self.use_image_source,
-                self.view_layer_name,
-            )
-        except ReliefError as e:
-            self.report({"ERROR"}, str(e))
-            return {"CANCELLED"}
-        except RuntimeError as e:
-            self.report({"ERROR"}, str(e))
-            return {"CANCELLED"}
-
-        try:
-            relief(self)
-        except ReliefError as e:
-            self.report({"ERROR"}, str(e))
-            return {"CANCELLED"}
-        return {"FINISHED"}
-
-    def draw(self, context):
-        """Draw the user interface for the bas relief settings.
-
-        This method constructs the layout for the bas relief settings in the
-        Blender user interface. It includes various properties and options that
-        allow users to configure the bas relief calculations, such as selecting
-        images, adjusting parameters, and setting justification options. The
-        layout is dynamically updated based on user selections, providing a
-        comprehensive interface for manipulating bas relief settings.
-
-        Args:
-            context (bpy.context): The context in which the UI is being drawn.
-
-        Returns:
-            None: This method does not return any value; it modifies the layout
-            directly.
-        """
-
-        layout = self.layout
-        layout.use_property_split = True
-        layout.use_property_decorate = False
-
-        scene = context.scene
-
-        box = layout.box()
-        col = box.column(align=True)
-        col.label(text="Source")
-        if self.use_image_source:
-            col.prop_search(self, "source_image_name", bpy.data, "images")
-        else:
-            col.prop_search(self, "view_layer_name", context.scene, "view_layers")
-        col.prop(self, "use_image_source")
-        col.prop(self, "depth_exponent")
-        col.prop(self, "advanced")
-
-        box = layout.box()
-        col = box.column(align=True)
-        col.label(text="Parameters")
-        col.prop(self, "bit_diameter", text="Ball End Diameter (mm)")
-        col.prop(self, "pass_per_radius")
-        col.prop(self, "widthmm", text="Desired Width (mm)")
-        col.prop(self, "heightmm", text="Desired Height (mm)")
-        col.prop(self, "thicknessmm", text="Thickness (mm)")
-
-        box = layout.box()
-        col = box.column(align=True)
-        col.label(text="Justification")
-        col.prop(self, "justifyx")
-        col.prop(self, "justifyy")
-        col.prop(self, "justifyz")
-
-        box = layout.box()
-        col = box.column(align=True)
-        col.label(text="Silhouette")
-        col.prop(self, "silhouette_threshold", text="Threshold")
-        col.prop(self, "recover_silhouettes")
-        if self.recover_silhouettes:
-            col.prop(self, "silhouette_scale", text="Scale")
-            if self.advanced:
-                col.prop(self, "silhouette_exponent", text="Square Exponent")
-        # layout.template_curve_mapping(self,'curva')
-        # layout.prop(self,'attenuation')
-
-        box = layout.box()
-        col = box.column(align=True)
-        col.label(text="Iterations")
-        if self.advanced:
-            col.prop(self, "smooth_iterations", text="Smooth")
-        col.prop(self, "vcycle_iterations", text="V-Cycle")
-        col.prop(self, "linbcg_iterations", text="LINBCG")
-
-        if self.advanced:
-            layout.prop(self, "min_gridsize")
-        layout.prop(self, "decimate_ratio")
-        layout.prop(self, "use_planar")
-
-        layout.prop(self, "gradient_scaling_mask_use")
-        if self.advanced:
-            if self.gradient_scaling_mask_use:
-                layout.prop_search(self, "gradient_scaling_mask_name", bpy.data, "images")
-            layout.prop(self, "detail_enhancement_use")
-            if self.detail_enhancement_use:
-                layout.prop(self, "detail_enhancement_amount")
-
-
-class ProblemAreas(bpy.types.Operator):
-    """Find Bas Relief Problem Areas"""
-
-    bl_idname = "scene.problemareas_bas_relief"
-    bl_label = "Problem Areas Bas Relief"
-    bl_options = {"REGISTER", "UNDO"}
-
-    processes = []
-
-    # @classmethod
-    # def poll(cls, context):
-    # 	return context.active_object is not None
-
-    def execute(self, context):
-        """Execute the operation related to the bas relief settings in the current
-        scene.
-
-        This method retrieves the current scene from the Blender context and
-        accesses the bas relief settings. It then calls the `problemAreas`
-        function to perform operations related to those settings. The method
-        concludes by returning a status indicating that the operation has
-        finished successfully.
-
-        Args:
-            context (bpy.context): The current Blender context, which provides access
-
-        Returns:
-            dict: A dictionary with a status key indicating the operation result,
-            specifically {'FINISHED'}.
-        """
-
-        s = bpy.context.scene
-        br = s.basreliefsettings
-        problem_areas(br)
-        return {"FINISHED"}
