+++ scripts/addons/cam/operators/curve_create_ops.py
@@ -30,6 +30,7 @@ from ..cam_chunk import (
     polygon_boolean,
     polygon_convex_hull,
 )
+
 from ..utilities.simple_utils import (
     remove_multiple,
     select_multiple,
