+++ scripts/addons/cam/operators/curve_equation_ops.py
