+++ scripts/addons/cam/utilities/geom_utils.py
@@ -1,4 +1,4 @@
-"""Fabex 'utils.py' © 2012 Vilem Novak
+"""Fabex 'geom_utils.py' © 2012 Vilem Novak
 
 Main functionality of Fabex.
 The functions here are called with operators defined in 'ops.py'
